Cache parsed value as a System.DateTime, unify unit tests
