+++ src/Hl7.Fhir.Support.Poco.Tests/Model/ModelTests.cs
@@ -6,9 +6,7 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-net-sdk/master/LICENSE
  */
 
-using FluentAssertions;
 using Hl7.Fhir.Model;
-using Hl7.Fhir.Utility;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
@@ -18,7 +16,7 @@ using System.Xml.Linq;
 namespace Hl7.Fhir.Tests.Model
 {
     [TestClass]
-    public partial class ModelTests
+    public class ModelTests
     {
         [TestMethod]
         public void ValidateElementAssertions()
@@ -152,37 +150,37 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void SimpleValueSupport()
         {
-            CapabilityStatement c = new CapabilityStatement();
+            Bundle b = new();
 
-            Assert.IsNull(c.Experimental);
-            c.Experimental = true;
-            Assert.IsTrue(c.Experimental.GetValueOrDefault());
-            Assert.IsNotNull(c.Experimental);
-            Assert.IsTrue(c.ExperimentalElement.Value.GetValueOrDefault());
+            Assert.IsNull(b.Total);
+            b.Total = 10;
+            Assert.AreEqual(10, b.Total.GetValueOrDefault());
+            Assert.IsNotNull(b.Total);
 
-            c.PublisherElement = new FhirString("Furore");
-            Assert.AreEqual("Furore", c.Publisher);
-            c.Publisher = null;
-            Assert.IsNull(c.PublisherElement);
-            c.Publisher = "Furore";
-            Assert.IsNotNull(c.PublisherElement);
+            OperationOutcome.IssueComponent ic = new();
 
-            c.Format = new string[] { "json", "xml" };
-            Assert.IsNotNull(c.FormatElement);
-            Assert.AreEqual(2, c.FormatElement.Count);
-            Assert.AreEqual("json", c.FormatElement.First().Value);
+            ic.DiagnosticsElement = new FhirString("Furore");
+            Assert.AreEqual("Furore", ic.Diagnostics);
+            ic.Diagnostics = null;
+            Assert.IsNull(ic.DiagnosticsElement);
+            ic.Diagnostics = "Furore";
+            Assert.IsNotNull(ic.DiagnosticsElement);
 
-            c.FormatElement = new List<Code>();
-            c.FormatElement.Add(new Code("csv"));
-            Assert.IsNotNull(c.Format);
-            Assert.AreEqual(1, c.Format.Count());
+            ic.Expression = new string[] { "json", "xml" };
+            Assert.IsNotNull(ic.ExpressionElement);
+            Assert.AreEqual(2, ic.ExpressionElement.Count);
+            Assert.AreEqual("json", ic.ExpressionElement.First().Value);
+
+            ic.ExpressionElement = new List<FhirString>();
+            ic.ExpressionElement.Add(new FhirString("csv"));
+            Assert.IsNotNull(ic.Expression);
+            Assert.AreEqual(1, ic.Expression.Count());
         }
-
 
         [TestMethod]
         public void ExtensionManagement()
         {
-            Patient p = new Patient();
+            TestPatient p = new();
             var u1 = "http://fhir.org/ext/ext-test";
             Assert.IsNull(p.GetExtension("http://fhir.org/ext/ext-test"));
 
@@ -244,9 +242,9 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void FindContainedResource()
         {
-            var cPat1 = new Patient() { Id = "pat1" };
-            var cPat2 = new Patient() { Id = "pat2" };
-            var pat = new Patient
+            var cPat1 = new TestPatient() { Id = "pat1" };
+            var cPat2 = new TestPatient() { Id = "pat2" };
+            var pat = new TestPatient
             {
                 Contained = new List<Resource> { cPat1, cPat2 }
             };
@@ -261,17 +259,18 @@ namespace Hl7.Fhir.Tests.Model
 
             Assert.AreEqual(pat, pat.FindContainedResource("#"));
 
-            var pat2 = new Patient();
+            var pat2 = new TestPatient();
             Assert.IsNull(pat2.FindContainedResource("#pat1"));
         }
+
 
         [TestMethod]
         [ExpectedException(typeof(ArgumentNullException))]
         public void FindContainedResourceExceptionExpected()
         {
-            var cPat1 = new Patient() { Id = "pat1" };
-            var cPat2 = new Patient() { Id = "pat2" };
-            var pat = new Patient
+            var cPat1 = new TestPatient() { Id = "pat1" };
+            var cPat2 = new TestPatient() { Id = "pat2" };
+            var pat = new TestPatient
             {
                 Contained = new List<Resource> { cPat1, cPat2 }
             };
@@ -282,12 +281,12 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void TestListDeepCopy()
         {
-            var x = new List<Patient>();
-            x.Add(new Patient());
-            x.Add(new Patient());
+            var x = new List<TestPatient>();
+            x.Add(new TestPatient());
+            x.Add(new TestPatient());
 
-            var y = new List<Patient>(x.DeepCopy());
-            Assert.IsTrue(x[0] is Patient);
+            var y = new List<TestPatient>(x.DeepCopy());
+            Assert.IsTrue(x[0] is TestPatient);
             Assert.AreNotEqual(x[0], y[0]);
             Assert.AreNotEqual(x[1], y[1]);
         }
@@ -296,45 +295,10 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void TestLazyCreatedLists()
         {
-            var p = new Patient();
-            p.Name.Add(new HumanName());
+            var p = new TestPatient();
+            p.Name.Add(new TestHumanName());
         }
 
-
-        [TestMethod]
-        public void TestModelInfoTypeSelectors()
-        {
-            Assert.IsTrue(ModelInfo.IsKnownResource("Patient"));
-            Assert.IsFalse(ModelInfo.IsKnownResource("Identifier"));
-            Assert.IsFalse(ModelInfo.IsKnownResource("code"));
-
-            Assert.IsFalse(ModelInfo.IsDataType("Patient"));
-            Assert.IsTrue(ModelInfo.IsDataType("Identifier"));
-            Assert.IsFalse(ModelInfo.IsDataType("code"));
-
-            Assert.IsFalse(ModelInfo.IsPrimitive("Patient"));
-            Assert.IsFalse(ModelInfo.IsPrimitive("Identifier"));
-            Assert.IsTrue(ModelInfo.IsPrimitive("code"));
-
-            Assert.IsTrue(ModelInfo.IsReference("Reference"));
-            Assert.IsFalse(ModelInfo.IsReference("Patient"));
-        }
-
-        [TestMethod]
-        public void TestFhirTypeToFhirTypeName()
-        {
-            var enumValues = Enum.GetValues(typeof(FHIRAllTypes));
-            for (int i = 0; i < enumValues.Length; i++)
-            {
-                var type = (FHIRAllTypes)i;
-                var typeName = ModelInfo.FhirTypeToFhirTypeName(type);
-                var type2 = ModelInfo.FhirTypeNameToFhirType(typeName);
-                Assert.IsTrue(type2.HasValue);
-                Assert.AreEqual(type, type2, String.Format("Failed: '{0}' != '{1}' ?!", type, type2));
-                var typeName2 = ModelInfo.FhirTypeToFhirTypeName(type2.Value);
-                Assert.AreEqual(typeName, typeName2, String.Format("Failed: '{0}' != '{1}' ?!", typeName, typeName2));
-            }
-        }
 
         [TestMethod]
         public void TestStringValueInterface()
@@ -383,55 +347,6 @@ namespace Hl7.Fhir.Tests.Model
 
 
         [TestMethod]
-        public void TestSubclassInfo()
-        {
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Resource, FHIRAllTypes.Patient));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.DomainResource, FHIRAllTypes.Patient));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Patient, FHIRAllTypes.Patient));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Observation, FHIRAllTypes.Patient));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Element, FHIRAllTypes.Patient));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Resource, FHIRAllTypes.Bundle));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.DomainResource, FHIRAllTypes.Bundle));
-
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Element, FHIRAllTypes.HumanName));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Element, FHIRAllTypes.Patient));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Element, FHIRAllTypes.Oid));
-
-            // FHIR: Canonical derives from Uri (not from String), but as gForge cofirmed Url and Canonical cannot be used as substitutes for Uri
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Uri, FHIRAllTypes.Canonical));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Uri, FHIRAllTypes.Url));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.String, FHIRAllTypes.Canonical));
-
-            // FHIR: Money derives from Element, not Quantity.
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(FHIRAllTypes.Quantity, FHIRAllTypes.Money));
-        }
-
-        [TestMethod]
-        public void TestSubclassInfoByType()
-        {
-            testTrue(typeof(Resource), typeof(Patient));
-            testTrue(typeof(DomainResource), typeof(Patient));
-            testTrue(typeof(Patient), typeof(Patient));
-            testFalse(typeof(Observation), typeof(Patient));
-            testFalse(typeof(Element), typeof(Patient));
-            testTrue(typeof(Resource), typeof(Bundle));
-            testFalse(typeof(DomainResource), typeof(Bundle));
-
-            testTrue(typeof(Element), typeof(HumanName));
-            testFalse(typeof(Element), typeof(Patient));
-            testTrue(typeof(Element), typeof(Oid));
-            testFalse(typeof(FhirString), typeof(Markdown));
-            testFalse(typeof(Integer), typeof(UnsignedInt));
-
-            static void testTrue(Type super, Type sub) =>
-                Assert.IsTrue(ModelInfo.IsInstanceTypeFor(super, sub));
-
-            static void testFalse(Type super, Type sub) =>
-                Assert.IsFalse(ModelInfo.IsInstanceTypeFor(super, sub));
-        }
-
-
-        [TestMethod]
         public void TestIntegerValueInterface()
         {
             INullableValue<int> iv = new Integer(null);
@@ -453,16 +368,16 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void TestChildren_EmptyPatient()
         {
-            var patient = new Patient();
+            var patient = new TestPatient();
             var children = patient.Children.ToArray();
             Base[] expected = { };
             Assert.IsTrue(expected.SequenceEqual(children));
         }
 
         [TestMethod]
-        public void TestChildren_EmptyTiming()
+        public void TestChildren_EmptyHumanName()
         {
-            var timing = new Timing();
+            var timing = new TestHumanName();
             var children = timing.Children.ToArray();
             Base[] expected = { };
             Assert.IsTrue(expected.SequenceEqual(children));
@@ -482,16 +397,16 @@ namespace Hl7.Fhir.Tests.Model
         [TestMethod]
         public void TestChildren_Patient()
         {
-            var patient = new Patient()
+            var patient = new TestPatient()
             {
                 Name =
                 {
-                    new HumanName()
+                    new TestHumanName()
                     {
                         Given = new string[] { "John" },
                         Family = "Doe"
                     },
-                     new HumanName()
+                     new TestHumanName()
                     {
                         Given = new string[] { "Alias" },
                         Family = "Alternate"
@@ -499,7 +414,7 @@ namespace Hl7.Fhir.Tests.Model
                 },
                 Address =
                 {
-                    new Address()
+                    new TestAddress()
                     {
                         City = "Amsterdam",
                         Line = new string[] { "Rokin" }
@@ -576,155 +491,6 @@ namespace Hl7.Fhir.Tests.Model
                 // address.Period
             };
             Assert.IsTrue(expected.SequenceEqual(children));
-        }
-
-        [TestMethod]
-        public void ParseFhirTypeName()
-        {
-            Assert.AreEqual(FHIRAllTypes.Markdown, ModelInfo.FhirTypeNameToFhirType("markdown"));
-            Assert.IsNull(ModelInfo.FhirTypeNameToFhirType("Markdown"));
-            Assert.AreEqual(FHIRAllTypes.Organization, ModelInfo.FhirTypeNameToFhirType("Organization"));
-        }
-
-        // [WMR 20181025] Issue #746
-        [TestMethod]
-        public void TestIsCoreModelTypeUri()
-        {
-            Assert.IsTrue(ModelInfo.IsCoreModelTypeUri(new Uri("http://hl7.org/fhir/StructureDefinition/Patient")));
-            Assert.IsTrue(ModelInfo.IsCoreModelTypeUri(new Uri("http://hl7.org/fhir/StructureDefinition/string")));
-
-            Assert.IsFalse(ModelInfo.IsCoreModelTypeUri(new Uri("http://example.org/fhir/StructureDefinition/Patient")));
-            Assert.IsFalse(ModelInfo.IsCoreModelTypeUri(new Uri("/StructureDefinition/Patient", UriKind.Relative)));
-            Assert.IsFalse(ModelInfo.IsCoreModelTypeUri(new Uri("Patient", UriKind.Relative)));
-        }
-
-
-        [TestMethod]
-        public void TestNonGeneratedHierarchy()
-        {
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(typeof(Quantity), typeof(Age)));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(typeof(DataType), typeof(Age)));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(typeof(Integer), typeof(UnsignedInt)));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(typeof(PrimitiveType), typeof(UnsignedInt)));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(typeof(FhirString), typeof(Code)));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(typeof(PrimitiveType), typeof(Code)));
-            Assert.IsFalse(ModelInfo.IsInstanceTypeFor(typeof(FhirUri), typeof(Uuid)));
-            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(typeof(PrimitiveType), typeof(Uuid)));
-        }
-
-        // [WMR 20190413] NEW
-
-        IEnumerable<Type> FhirCsTypes => ModelInfo.FhirCsTypeToString.Keys;
-
-        [TestMethod]
-        public void TestIsConformanceResource()
-        {
-            // Verify that ModelInfo.IsConformanceResource overloads returns true for all types that implement IConformanceResource
-            foreach (var type in FhirCsTypes)
-            {
-                //var supportsInterface = typeof(IConformanceResource).IsAssignableFrom(type);
-                var supportsInterface = type.CanBeTreatedAsType(typeof(IConformanceResource));
-                Assert.AreEqual(supportsInterface, ModelInfo.IsConformanceResource(type));
-                var typeName = ModelInfo.GetFhirTypeNameForType(type);
-                Assert.AreEqual(supportsInterface, ModelInfo.IsConformanceResource(typeName));
-                var typeFlag = ModelInfo.FhirTypeNameToFhirType(typeName);
-                Assert.AreEqual(supportsInterface, ModelInfo.IsConformanceResource(typeFlag));
-            }
-        }
-
-        [TestMethod]
-        public void TestIsPrimitive()
-        {
-            // Verify that ModelInfo.IsPrimitive overloads returns true for all types derived from Primitive
-            foreach (var type in FhirCsTypes)
-            {
-                var isPrimitive = type.CanBeTreatedAsType(typeof(PrimitiveType));
-                var typeName = ModelInfo.GetFhirTypeNameForType(type);
-                Assert.IsNotNull(typeName);
-                Assert.AreEqual(isPrimitive, ModelInfo.IsPrimitive(type));
-                Assert.AreEqual(isPrimitive, ModelInfo.IsPrimitive(typeName));
-            }
-        }
-
-        [TestMethod]
-        public void TestIsDataType()
-        {
-            // Verify that ModelInfo.IsDataType returns true for all types derived from Element but not from Primitive
-            foreach (var type in FhirCsTypes)
-            {
-                if (type == typeof(Base)) continue;
-
-                var isDataType = type.CanBeTreatedAsType(typeof(Element)) && !type.CanBeTreatedAsType(typeof(PrimitiveType)); var typeName = ModelInfo.GetFhirTypeNameForType(type);
-                Assert.IsNotNull(typeName);
-                Assert.AreEqual(isDataType, ModelInfo.IsDataType(type), type.Name);
-                Assert.AreEqual(isDataType, ModelInfo.IsDataType(typeName));
-            }
-        }
-
-        [TestMethod]
-        public void TestIsReference()
-        {
-            // Verify that ModelInfo.IsReference overloads returns true for type (Resource)Reference
-            foreach (var type in FhirCsTypes)
-            {
-                var isReference = type == typeof(ResourceReference);
-                var typeName = ModelInfo.GetFhirTypeNameForType(type);
-                Assert.IsNotNull(typeName);
-                Assert.AreEqual(isReference, ModelInfo.IsReference(type));
-                Assert.AreEqual(isReference, ModelInfo.IsReference(typeName));
-            }
-        }
-
-        [TestMethod]
-        public void TestTypeHierarchy()
-        {
-            // Verify ModelInfo methods are in agreement with the actual type hierarchy
-            // - ModelInfo.IsInstanceTypeFor
-            // - ModelInfo.IsCoreSuperType
-            var types = ModelInfo.FhirCsTypeToString.Keys;
-            foreach (var type in types)
-            {
-                Assert.IsTrue(ModelInfo.IsCoreModelType(type));
-                var typeName = ModelInfo.GetFhirTypeNameForType(type);
-                Assert.IsNotNull(typeName);
-                Assert.IsTrue(ModelInfo.IsCoreModelType(typeName));
-
-                if (!ModelInfo.IsCoreSuperType(type))
-                {
-                    var baseType = type.BaseType;
-                    while (!ModelInfo.IsCoreSuperType(baseType))
-                    {
-                        // Skip intermediate abstract types, e.g. Primitive<T>
-                        if (ModelInfo.IsCoreModelType(baseType))
-                        {
-                            var baseTypeName = ModelInfo.GetFhirTypeNameForType(baseType);
-                            Assert.IsNotNull(baseTypeName);
-
-                            Assert.IsTrue(ModelInfo.IsInstanceTypeFor(baseTypeName, typeName));
-                        }
-                        baseType = baseType.BaseType;
-                    }
-                }
-            }
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void TestCheckMinorVersionCompatibilityWithNull()
-        {
-            ModelInfo.CheckMinorVersionCompatibility(null);
-        }
-
-        [TestMethod]
-        public void TestHumanNameFluentInitializers()
-        {
-            var pat = new Patient();
-            pat.Name.Add(new HumanName().WithPrefix("Mr.").WithGiven("Henry").AndFamily("Levin").WithSuffix("The 7th"));
-
-            pat.Name.FirstOrDefault().Prefix.Should().ContainSingle("Mr.");
-            pat.Name.FirstOrDefault().Given.Should().ContainSingle("Henry");
-            pat.Name.FirstOrDefault().Family.Should().Be("Levin");
-            pat.Name.FirstOrDefault().Suffix.Should().ContainSingle("The 7th");
         }
     }
 }
